requirements.txt
Flask==3.1.1
eventlet==0.40.0
flask-socketio==5.5.1
Pillow==11.2.1
python-dotenv==1.0.0
qrcode==8.2
flask
requests
flask-socketio
python-dotenv
qrcode
Pillow